backend/requirements.txt
# Biomni Backend Requirements

# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
sse-starlette==1.6.5

# LLM and AI
openai==1.3.0
langchain==0.1.0
langchain-openai==0.0.2
dspy-ai==2.4.0

# Scientific computing
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1
scikit-learn==1.3.0
matplotlib==3.7.1

# Bioinformatics
biopython==1.81
rdkit==2023.3.2

# Data processing
aiofiles==23.2.1
python-multipart==0.0.6
Pillow==10.0.0

# Database
pymongo==4.5.0
redis==5.0.0

# Utilities
tqdm==4.66.1
rich==13.7.0
pyyaml==6.0.1
jsonschema==4.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.0